+++ test/create_update_str_cow/schema.rs
@@ -1,0 +1,9 @@
+diesel::table! {
+    todos (text) {
+        text -> Text,
+        text_nullable -> Nullable<Text>,
+        #[max_length = 255]
+        varchar -> Varchar,
+        varchar_nullable -> Nullable<Varchar>,
+    }
+}