+++ test/create_update_str_str/models/todos/generated.rs
@@ -1,0 +1,92 @@
+/* @generated and managed by dsync */
+
+use crate::diesel::*;
+use crate::schema::*;
+use diesel::QueryResult;
+use serde::{Deserialize, Serialize};
+
+
+type ConnectionType = diesel::r2d2::PooledConnection<diesel::r2d2::ConnectionManager<diesel::PgConnection>>;
+
+#[derive(Debug, Serialize, Deserialize, Clone, Queryable, Insertable, AsChangeset, Selectable)]
+#[diesel(table_name=todos, primary_key(text))]
+pub struct Todos {
+    pub text: String,
+    pub text_nullable: Option<String>,
+    pub varchar: String,
+    pub varchar_nullable: Option<String>,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone, Queryable, Insertable, AsChangeset)]
+#[diesel(table_name=todos)]
+pub struct CreateTodos<'a> {
+    pub text: &'a str,
+    pub text_nullable: Option<&'a str>,
+    pub varchar: &'a str,
+    pub varchar_nullable: Option<&'a str>,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone, Queryable, Insertable, AsChangeset, Default)]
+#[diesel(table_name=todos)]
+pub struct UpdateTodos<'a> {
+    pub text_nullable: Option<Option<&'a str>>,
+    pub varchar: Option<&'a str>,
+    pub varchar_nullable: Option<Option<&'a str>>,
+}
+
+
+#[derive(Debug, Serialize)]
+pub struct PaginationResult<T> {
+    pub items: Vec<T>,
+    pub total_items: i64,
+    /// 0-based index
+    pub page: i64,
+    pub page_size: i64,
+    pub num_pages: i64,
+}
+
+impl Todos {
+
+    pub fn create(db: &mut ConnectionType, item: &CreateTodos) -> QueryResult<Self> {
+        use crate::schema::todos::dsl::*;
+
+        insert_into(todos).values(item).get_result::<Self>(db)
+    }
+
+    pub fn read(db: &mut ConnectionType, param_text: String) -> QueryResult<Self> {
+        use crate::schema::todos::dsl::*;
+
+        todos.filter(text.eq(param_text)).first::<Self>(db)
+    }
+
+    /// Paginates through the table where page is a 0-based index (i.e. page 0 is the first page)
+    pub fn paginate(db: &mut ConnectionType, page: i64, page_size: i64) -> QueryResult<PaginationResult<Self>> {
+        use crate::schema::todos::dsl::*;
+
+        let page_size = if page_size < 1 { 1 } else { page_size };
+        let total_items = todos.count().get_result(db)?;
+        let items = todos.limit(page_size).offset(page * page_size).load::<Self>(db)?;
+
+        Ok(PaginationResult {
+            items,
+            total_items,
+            page,
+            page_size,
+            /* ceiling division of integers */
+            num_pages: total_items / page_size + i64::from(total_items % page_size != 0)
+        })
+    }
+
+    pub fn update(db: &mut ConnectionType, param_text: String, item: &UpdateTodos) -> QueryResult<Self> {
+        use crate::schema::todos::dsl::*;
+
+        diesel::update(todos.filter(text.eq(param_text))).set(item).get_result(db)
+    }
+
+    pub fn delete(db: &mut ConnectionType, param_text: String) -> QueryResult<usize> {
+        use crate::schema::todos::dsl::*;
+
+        diesel::delete(todos.filter(text.eq(param_text))).execute(db)
+    }
+
+}